# add multiplication with vectors method