--- a/Project.toml
+++ b/Project.toml
@@ -8,7 +8,7 @@ FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 
 [compat]
-FillArrays = "0.6, 0.7"
+FillArrays = "0.6, 0.7, 0.8"
 julia = "1"
 
 [extras]