This seems to fix an oversight. Previously, there was no specialized `*(::BlockDiagonal,::AbstractVector)` method, such that this combination was handled by some generic `LinearAlgebra` method. Not surprisingly, this yields a giant performance gain.

I have kept the changes simple and minimal. There is quite some code duplication with the corresponding `*(::BlockDiagonal,::AbstractMatrix)` method, which can be reduced by means of a helper function and using `selectdim(y, 1, st:ed)` if desired.