@nickrobinson251 2019-12-09
🎉 Thanks a lot! LGTM

> There is quite some code duplication with the corresponding *(::BlockDiagonal,::AbstractMatrix) method, which can be reduced by means of a helper function and using selectdim(y, 1, st:ed) if desired.

That'd be a great follow-up PR, if you have time for that? 

For this PR, would you mind posting a simple benchmark giving a quick illustration of the speed-up (it'd be helpful when one day making proper benchmarks for the package #19)

And please bump the package version -- will tag a new release straight away!