@dkarrasch 2019-12-10
Here is a quick benchmark, slightly adopted from the test suite:
```julia
using BlockDiagonals, BenchmarkTools
N1, N2, N3 = 30, 40, 50
N = N1 + N2 + N3
b1 = BlockDiagonal([rand(N1, N1), rand(N2, N2), rand(N3, N3)])
a = rand(N)
@benchmark $b1*$a
```
Master:
```julia
BenchmarkTools.Trial: 
  memory estimate:  4.41 MiB
  allocs estimate:  43921
  --------------
  minimum time:     1.702 ms (0.00% GC)
  median time:      1.852 ms (0.00% GC)
  mean time:        2.195 ms (11.26% GC)
  maximum time:     7.055 ms (44.17% GC)
  --------------
  samples:          2270
  evals/sample:     1
```
This PR:
```julia
BenchmarkTools.Trial: 
  memory estimate:  2.63 KiB
  allocs estimate:  9
  --------------
  minimum time:     1.121 μs (0.00% GC)
  median time:      1.227 μs (0.00% GC)
  mean time:        1.633 μs (4.68% GC)
  maximum time:     180.318 μs (97.38% GC)
  --------------
  samples:          10000
  evals/sample:     10
```